skip method takse long as an argument, not int
this required implementing iterative skip in random access adapter as
random access file's skip method accepts int


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActive_DataSpaces@13261 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/Extra/org/objectweb/proactive/extra/vfsprovider/server/InputStreamAdapter.java
@@ -42,7 +42,7 @@ public class InputStreamAdapter implements Stream {
         throw new WrongStreamTypeException();
     }
 
-    public long skip(int bytes) throws IOException {
+    public long skip(long bytes) throws IOException {
         return adaptee.skip(bytes);
     }
